requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==42.0.0 